order.js에서 값받아오는 경로 수정
