+++ src/pages/ReviewAdd.jsx
@@ -104,10 +104,10 @@ const ReviewAdd = () => {
   // mypage의 배송완료 상품에서 받아온 프롭
   const location = useLocation();
   const [data, setData] = useState("");
+  const [checkId, setCheckId] = useState("");
+  console.log(data);
 
-  // console.log(location);
-
-  // 서브밋 함수
+  // 글자수와 별점을 선택하게 하는 함수
   const reviewSumbit = (e) => {
     e.preventDefault();
     if (comment?.length < 10) {
@@ -119,7 +119,7 @@ const ReviewAdd = () => {
     }
 
     const newReview = {
-      ...data,
+      // ...data,
       thumbnail: data.thumbnail,
       img,
       userID,
@@ -132,7 +132,10 @@ const ReviewAdd = () => {
       productID: data.productID,
     };
 
-    if (userID === location.state.user) {
+    // 로그인유저가 작성한 유저가 같다면 작성내용을 input리듀서로 디스패치함
+    // 코멘트 내용이 다르다면 edit리듀서로 디스패치함
+    // 리뷰등록한 뒤 맞는 productID로 navigate됨
+    if (userID === checkId) {
       dispatch(inputReview(newReview));
     } else if (location.state.comment !== comment) {
       dispatch(editReview(newReview));
@@ -143,11 +146,12 @@ const ReviewAdd = () => {
 
   // 마이페이지에서 값을 잘 받아오고 있으면 (주소창에 mypaye/review로 접근하는 등 편법 방지) 에러페이지를 출력함
   useEffect(() => {
-    if (!location.state) {
+    if (!location.state.order) {
       alert("잘못된 경로로 접근하였습니다.");
       navigate("/notfound");
     } else {
-      setData(location.state);
+      setData(location.state.order);
+      setCheckId(location.state.userId);
     }
   }, []);
 
@@ -167,11 +171,7 @@ const ReviewAdd = () => {
       <form onSubmit={reviewSumbit} className="review-submit-form">
         <section className="review-form-product">
           {/* 상품이미지 box*/}
-          <img
-            src={getImgPath(data)}
-            alt="No Image"
-            style={{ width: "120px", height: "120px" }}
-          />
+          <img src={getImgPath(data)} alt="No Image" style={{ width: "120px", height: "120px" }} />
 
           {/* 상품옵션 box */}
           <div>
@@ -236,27 +236,18 @@ const ReviewAdd = () => {
                 <strong>사진 첨부하기</strong>
               </span>
             </MyButton>
-            
+
             {/* 미리보기 사진 전달공간 */}
             {/* 전달받은 이미지가 있으면 영역출력, 없으면 빈 div */}
             {img ? (
               <div>
-                <img
-                  src={img}
-                  alt=""
-                  style={{ width: "180px", height: "240px", marginTop: "10px" }}
-                />
+                <img src={img} alt="" style={{ width: "180px", height: "240px", marginTop: "10px" }} />
               </div>
             ) : (
               <div></div>
             )}
             {/* 사진첨부 모달창*/}
-            <Modal
-              open={modalOpen}
-              onClose={handleClose}
-              aria-labelledby="modal-modal-title"
-              aria-describedby="modal-modal-description"
-            >
+            <Modal open={modalOpen} onClose={handleClose} aria-labelledby="modal-modal-title" aria-describedby="modal-modal-description">
               <Box sx={style} className="review-modal">
                 <header>
                   <p>사진 첨부</p>
@@ -304,10 +295,7 @@ const ReviewAdd = () => {
                     border: "solid 1px gray",
                   }}
                 >
-                  <FontAwesomeIcon
-                    icon={faCamera}
-                    style={{ margin: "0 5px" }}
-                  />
+                  <FontAwesomeIcon icon={faCamera} style={{ margin: "0 5px" }} />
                   <label htmlFor="imageInput">사진추가</label>
                 </MyButton>
