+++ src/redux/reducers/reviewInputReducer.js
@@ -21,7 +21,7 @@ const reviewSlice = createSlice({
     inputReview: (state, action) => {
       // console.log(action.payload.date);
       const newReview = {
-        reviewID: 1,
+        reviewID: state.reviewID,
         // 상품ID 불러오기
         productID: action.payload.productID,
         // 리뷰 첨부 이미지 불러오기
@@ -45,23 +45,20 @@ const reviewSlice = createSlice({
         // // 작성 날짜 불러오기
         date: getDate(),
       };
+      state.reviewID++;
       const newReviewlist = state.reviewlist.concat(newReview);
       state.reviewlist = newReviewlist;
     },
     // 리뷰 삭제하기
     deleteReview: (state, action) => {
       // reviewInput배열에서 reiviewID가 파라미터로 일치하지 않는 원소만 추출해서 새로운 배열을 만듦
-      const newReviewlist = state.reviewlist.filter(
-        (review) => review.userID == action.payload
-      );
+      const newReviewlist = state.reviewlist.filter((review) => review.userID == action.payload);
       state.reviewlist = newReviewlist;
     },
     // 리뷰 수정하기
     editReview: (state, action) => {
       state.reviewlist = state.reviewlist.map((review) => {
-        return action.payload.userID === review.userID
-          ? action.payload
-          : review;
+        return action.payload.userID === review.userID ? action.payload : review;
       });
     },
   },
